Fix next bunch of warnings

+++ sdk/core/core/src/az_hex_private.h
@@ -19,7 +19,7 @@ enum
  */
 AZ_NODISCARD AZ_INLINE uint8_t _az_number_to_upper_hex(uint8_t number)
 {
-  return number + (number < 10 ? '0' : _az_HEX_UPPER_OFFSET);
+  return (uint8_t)(number + (number < 10 ? '0' : _az_HEX_UPPER_OFFSET));
 }
 
 #include <_az_cfg_suffix.h>
